// src/db/enums/ticket.enum.ts
export enum TicketStatus {
  open = 'open',
  resolved = 'resolved',
}

export enum TicketType {
  managementReport = 'managementReport',
  registrationAddressChange = 'registrationAddressChange',
  strikeOff = 'strikeOff',
}

export enum TicketCategory {
  accounting = 'accounting',
  corporate = 'corporate',
  management = 'management',
}